fix: resolve ambiguous foreign key in User.personas relationship
Specify foreign_keys parameter in User.personas relationship to resolve
AmbiguousForeignKeysError. The Persona model has two foreign keys to User
(creator_id and original_creator_id), causing SQLAlchemy to be unable to
determine which one to use for the relationship. Explicitly setting
foreign_keys="Persona.creator_id" resolves this issue.

This fixes the 500 error when authenticating with Firebase/Google Sign-In.